+++ lib/generators/cms/attribute/templates/migrations/html_migration.rb
@@ -1,0 +1,5 @@
+class Create<%= class_name %>Attribute < ::RailsConnector::Migration
+  def up
+    create_attribute(name: '<%= file_name %>', type: '<%= type %>', title: '<%= title %>')
+  end
+end