+++ lib/generators/cms/kickstart/kickstart_generator.rb
@@ -78,7 +78,7 @@ module Cms
       def create_deploy_hooks
         create_file('deploy/after_restart.rb')
         create_file('deploy/before_symlink.rb')
-        create_file('deploy/before_migrate.rb')
+        prepend_file('deploy/before_migrate.rb', 'deploy/before_migrate.rb')
         empty_directory('deploy/templates')
       end
 
@@ -151,8 +151,9 @@ module Cms
 
         Rails::Generators.invoke('cms:attribute', ['source', '--type=linklist', '--title=Source'])
         Rails::Generators.invoke('cms:attribute', ['caption', '--type=string', '--title=Caption'])
+        Rails::Generators.invoke('cms:attribute', ['link_to', '--type=linklist', '--title=Link'])
         Rails::Generators.invoke('cms:model', ['BoxText', '--title=Box: Text', '--attributes=sort_key'])
-        Rails::Generators.invoke('cms:model', ['BoxImage', '--title=Box: Image', '--attributes=source', 'caption', 'sort_key'])
+        Rails::Generators.invoke('cms:model', ['BoxImage', '--title=Box: Image', '--attributes=source', 'caption', 'link_to', 'sort_key'])
 
         Rails::Generators.invoke('cms:attribute', ['redirect_link', '--type=linklist', '--title=Redirect Link', '--max-size=1'])
         Rails::Generators.invoke('cms:model', ['Redirect', '--title=Redirect', '--attributes=sort_key', 'redirect_link', 'show_in_navigation'])
@@ -167,6 +168,7 @@ module Cms
         directory('lib')
         directory('config')
         directory('deploy')
+        directory('spec')
       end
 
       def add_initital_components