+++ lib/generators/cms/kickstart/templates/deploy/before_migrate.rb
@@ -5,5 +5,3 @@ shared_asset_path = "#{new_resource.shared_path}/assets"
 run("mkdir -p #{shared_asset_path}")
 run("rm -rf #{asset_path}")
 run("cd #{public_path} && ln -s #{shared_asset_path}")
-
-run('bundle exec rake assets:precompile RAILS_ENV=production')